- EMAR-66: API for signing the new orders and adding to patients orders Emar.Api\Startup.cs Emar.Api\Controllers\CartOrdersController.cs Emar.Core\Carts\Model\Mappings\CartOrderMapper.cs Emar.Core\Carts\Model\CartOrderAdministrationDto.cs Emar.Core\Carts\Model\CartOrderDto.cs Emar.Core\Carts\Repository\CartOrderRepository.cs Emar.Core\Carts\Repository\ICartOrderRepository.cs Emar.Core\Carts\Service\CartOrderService.cs Emar.Core\Carts\Service\ICartOrderService.cs Emar.Core\Helpers\PropertyMappingService.cs Emar.Core\ResourceParameters\BaseResourceParameters.cs Emar.Data\Entities\CartOrderAdministration.cs Emar.Data\Entities\PatientCartOrder.cs Emar.Data\EmarContext.cs
- No ticket: Accommodate db changes
Emar.Api\Controllers\PatientsController.cs
Emar.Core\Medications\Model\MedicationRouteDto.cs
Emar.Core\Orders\Model\Mappings\OrderMapper.cs
Emar.Core\Orders\Model\PatientOrderDto.cs
Emar.Core\Orders\Repository\OrderRepository.cs
Emar.Core\Patients\Model\Mappings\PatientMapper.cs
Emar.Core\Patients\Repository\PatientRepository.cs
Emar.Core\ResourceParameters\PatientsResourceParameters .cs
Emar.Core\Sites\Model\Mappings\SiteMapper.cs
Emar.Core\Sites\Model\SiteDto.cs
Emar.Core\Users\Model\Mappings\UserMapper.cs
Emar.Core\Users\Model\UserDto.cs
Emar.Data\EmarContext.cs
Emar.Data\Entities\Action.cs
Emar.Data\Entities\GroupListItems.cs
Emar.Data\Entities\MedicationRoute.cs
Emar.Data\Entities\OrderAdministration.cs
Emar.Data\Entities\OrderEvent.cs
Emar.Data\Entities\Patient.cs
Emar.Data\Entities\PatientOrder.cs
Emar.Data\Entities\Site.cs
Emar.Data\Entities\User.cs

- No ticket: Add Medication Route in patient order record
Emar.Core\Orders\Model\Mappings\OrderMapper.cs

- No ticket: Trim extra spaces in all string values from the database
Emar.Core\Orders\Model\OrderBase.cs
Emar.Core\Patients\Model\PatientDto.cs
Emar.Core\Users\Model\UserDto.cs

- No ticket: Add MedicalRecordNumber in patient record
Emar.Core\Patients\Model\Mappings\PatientMapper.cs

- No ticket: Add Urgency in patient record
Emar.Core\Patients\Model\Mappings\PatientMapper.cs

- No ticket: Add the first name, middle name, last name and suffix in user record
Emar.Api\Controllers\UsersController.cs
Emar.Core\Users\Model\Mappings\UserMapper.cs
Emar.Core\Users\Model\UserDto.cs
Emar.Data\Entities\User.cs

- No ticket: Accept a string instead of int external id when calling for a user record
Emar.Api\Controllers\UsersController.cs
Emar.Core\Users\Repository\IUserRepository.cs
Emar.Core\Users\Repository\UserRepository.cs


git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C234

+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Core/ResourceParameters/PatientsResourceParameters .cs
@@ -17,7 +17,7 @@
         /// <summary>
         /// Site (facility) identifier to restrict the list of returned patients to.
         /// </summary>
-        public short? SiteId { get; set; }
+        public int? SiteId { get; set; }
         /// <summary>
         /// Department code to restrict the list of returned patients to.
         /// </summary>
